refactor: Move analyticGroupoid to a new AnalyticManifold.lean
I'm about to rewrite the definition of `analyticGroupoid` and define
`AnalyticManifold` on top of it, in
https://github.com/leanprover-community/mathlib4/pull/10853. @grunweg
suggested splitting out the move of `analyticGroupoid` into a separate
commit, so that the diff is easier. It used to be in
`SmoothManifoldWithCorners.lean`, but this makes no sense as nothing
downstream of that file uses `analyticGroupoid`.