feat(Topology): some lemmas about continuity of maps out of one point compactifications (#13355)
We prove that a map out of the one point compactification of `X` is continuous if and only if its restriction to `X` satisfies the two properties
1) is continuous
2) is continuous "at infinity", namely that it tends to the value at infinity for the `coclosedCompact` filter on `X`

We deduce that continuous functions out of the one point compactification of the natural numbers correspond to convergent sequences. 

We also prove that the one point compactification of a discrete space is totally separated. 

Co-authored-by: Jireh Loreaux



Co-authored-by: faenuccio <filippo.nuccio@univ-st-etienne.fr>