feat: Order-connected sets in `ℝⁿ` are null-measurable (#13633)
Prove that the frontier of an order-connected set in `ℝⁿ` (with the `∞`-metric, but it doesn't actually matter) has measure zero. As a corollary, antichains in `ℝⁿ` have measure zero.

This is not so trivial as one might think. The proof Kexing and I came up with involves the Lebesgue density theorem.

Partially forward-port leanprover-community/mathlib#16976