feat(Data/Set/Function): extending bijectivity to supersets (#13692)
the lemmas `exists_subset_bijOn` and similar show that for a function `f : α → β` and a set `t : Set β` contained in the range of `f`, we can choose a set `s` that `f` maps bijectively to `t`. We add a few lemmas stating that, under appropriate hypotheses, we can choose such an `s` to contain and/or be contained in prescribed sets in `α`.