Add constructor `RootPairing.mk'`
This shows that subject to the hypotheses:
 * finite number of roots / coroots
 * characteristic zero
 * no torsion
we do not need to demand that the reflections and coreflections induce the
same family of permuations on roots / coroots since this follows automatically.
This shows that our definition does indeed strictly generalise root datum (as
well as providing a useful constructor for users).

Also included are various minor style changes and API improvements.