feat: differentials as an object in ModuleCat (#14030)
In this PR, given a morphism `f : A ⟶ B` in the category `CommRingCat`, and `M : ModuleCat B`, we define the type `M.Derivation f` of derivations with values in `M` relative to `f`. We also construct the module of differentials `CommRingCat.KaehlerDifferential f : ModuleCat B` and the corresponding derivation. (These are basically "bundled" versions of the unbundled constructions already in mathlib.)



Co-authored-by: Joël Riou <37772949+joelriou@users.noreply.github.com>