feat(AlgebraicGeometry/ProjectiveSpectrum/Scheme): finish the `Proj` construction (#12371)
This PR finishes the `Proj` construction for any ring graded by natural numbers:
If $A$ is an $R$ algebra with 𝒜 as its grading i.e. 𝒜 i is an $R$-submodule of $A$ , then `AlgebraicGeometry.Proj 𝒜` is the scheme whose underlying topological space is the collection of homogeneous relevant prime ideals with the Zariski topology; whose sheaf of rings is the collection of functions that are locally homogeneous fractions. We prove that for each `f : A` with positive degree, `Proj | D(f)` (`Proj` as locally ringed space restricted to the basic open set around `f`) is isomorphic to `Spec A^0_f` (prime spectrum of the homogeneous localization of `A` at `f`).

The isomorphism is constructed as the following:
by the Gamma-Spec adjunction, it is sufficient to construct a ring map `A⁰_f → Γ(Proj, pbo f)` from the ring of homogeneous localization of `A` away from `f` to the local sections of structure sheaf of projective spectrum on the basic open set around `f`.

The map `A⁰_f → Γ(Proj, pbo f)` is defined by sending `s ∈ A⁰_f` to the section `x ↦ s` on `pbo f`.

Then we show that the map `Proj | D(f) -> Spec A⁰_f` induces an isomorphism on stalk level, thus is an isomorphism.

Co-authored-by: Andrew Yang <the.erd.one@gmail.com>



Co-authored-by: zjj <j.jj@j.jj>
Co-authored-by: Andrew Yang <the.erd.one@gmail.com>
Co-authored-by: Andrew Yang <36414270+erdOne@users.noreply.github.com>