feat: small types of morphisms in the localized category (#13639)
Given a class of morphisms `W : MorphismProperty C`, and two objects `X` and `Y`, we introduce a type-class `HasSmallLocalizedHom.{w} W X Y` which expresses that in the localized category with respect to `W`, the type of morphisms from `X` to `Y` is `w`-small for a certain universe `w`. Under this assumption, we define `SmallHom.{w} W X Y : Type w` as the shrunk type. For any localization functor `L : C ⥤ D` for `W`, we provide a bijection `SmallHom.equiv.{w} W L : SmallHom.{w} W X Y ≃ (L.obj X ⟶ L.obj Y)` that is compatible with the composition of morphisms.

This shall be used in the refactor of `Ext`-groups in abelian categories.



Co-authored-by: Joël Riou <37772949+joelriou@users.noreply.github.com>