feat(ContinuousFunctionalCalculus): add several lemmas involving the CFC and `algebraMap` (#14065)
This PR adds several lemmas about the interaction between the (non-unital) CFC and `algebraMap`. Several lemmas require some sort of nontriviality statement for the CFC (i.e. the predicate can't be false everywhere), I just stated it as an `hp : p 0` hypothesis in the lemma statements; it seems like the most convenient way in practice and probably the easiest to automate.

Co-authored-by: Jireh Loreaux <loreaujy@gmail.com>