feat (Algebra/Order/AddTorsor) : OrderedVAdd, OrderedCancelVAdd (#11979)
This PR introduces prop-valued mixins for vector addition operations that preserve or reflect order, and provides instances relating it to ordered commutative monoids and to covariant/contravariant classes.  In a future PR, we expect to use this to generalize Hahn module structures, for example giving an `R((z))`-module structure to `z^s V((z))` for `V` an `R`-module.