fix(nlinarith): make `findSquares` deterministic
Previously the order in which the assumptions were added was determined by `Expr` hashes,
which while deterministic are not part of the user's mental model.

This matters (see [Zulip](https://leanprover.zulipchat.com/#narrow/stream/287929-mathlib4/topic/nlinarith.20is.20affected.20by.20.60example.60s/near/449994710)) because it then affects the order in which metavariables are unified.