rename: `no_lost_declarations` to `declarations_diff` and `short` as default option (#14105)
This PR renames the `no_lost_declarations` script to `declarations_diff`, to better describe the effect of the script.  It was originally intended as a script to make sure that "moving" PRs did not leave out any declaration, but is now used as a measure of the diff in declarations between PRs and master.

The other change it does is to make the `short` option the default, as this is the most commonly used.

|Old|New|
|-|-|
|`./scripts/no_lost_declarations.sh short`|`./scripts/declarations_diff.sh`|
|`./scripts/no_lost_declarations.sh`|`./scripts/declarations_diff.sh long`|

Mostly, these changes are invisible, as this script is used almost exclusively by CI, but the intention is that the change conveys better what script does to a human.