feat(Algebra/Category/ModuleCat/{Presheaf, Sheaf}): forgetful functor from sheaves of modules over sheaf of ring R to sheaves of R(X)-module when X is initial (#14254)
Let $R : C^{op} \to Ring$ be a sheaf of rings and $M$ a sheaf of $R$-modules. Then $M$ is naturally a sheaf of $R(\top)$-modules where $\top$ is an initial object.

In the case of a sheafed space $X$, for any $x \in X$ and $M$ a sheaf of $\mathcal{O} _X$-modules, then $M$ is also a sheaf of $\Gamma(\mathcal{O} _X, X)$-modules; thus we can take stalks to naturally have that $M_x$ is a $\Gamma(\mathcal{O} _X, X)$-module. Thus we can use this construction to define $M_x$ as a module over $\mathcal{O} _{X, x}$.

Co-authored-by: Sophie Morel <sophie.morel@ens-lyon.fr>