fix: `maintainer merge` messages only once (#14830)
Testing the doubled `maintainer merge` messages.

The conclusion seems to be that `edited` triggers the action also when a comment is submitted as part of a review.  The change proposed here likely means that editing a message to make it contain `maintainer merge` will not trigger a `maintainer merge` message, but will not double `maintainer merge` messages that are sent from a review that also adds comments.