feat: define n-manifolds and closed manifolds
Define structures `NManifold` and `ClosedManifold` for n-dimensional resp. closed smooth manifolds (i.e., compact without boundary).
These are useful in their own right, and needed to even state the basic definition in bordism theory.

To show these definitions are workable, we state some standard examples:
- Euclidean space is an n-manifold
- the n-sphere is a closed n-manifold
- the two-torus is a closed 2-manifold

Defining n-manifolds posed a design question, whether to have the n-dimensionality as a typeclass assumption or a proof-valued field.
The former could avoid unnecessary boilerplate, say when the dimension is ambient and always fixed.
However, encoding the dimension in a typeclass is inflexible for dimension-changing constructions. Three examples:
- The product of an n- and an m-manifold is an n+m-manifold: I would prefer that a 2+3 and a 4+1-manifold are the same as a 5-manifold; the typeclass system and addition don't mix well.
- Bordism theory requires such equalities: a cobordism between two 4-manifolds is a certain 5-manifold;
the general definition would state this as a 4+1-manifold, which needs to be unified with 5.
- Surgery between manifolds involves k-handles, the product of a k- and an n-k-manifold (for all k): this uses quite extensively that n=k+(n-k) for all k.

For these reasons, I have decided for the latter option.

I am open to also defining a class IsNManifold and adding an instance SmoothNManifold -> IsNManifold.