fix: parse header in `minImports` to improve reports (#15081)
When reporting the summary of import changes, the `minImports` linter was using a text-based heuristic to determine the imports.  Kyle explained how to get exact syntax information for the header and this PR uses that for error reporting.

Note that there is no change to the linter, other than in how the report is printed.