chore: replace forall intro in UnsignedHahn (#15205)
This file (and specifically [MeasureTheory.hahn_decomposition](https://leanprover-community.github.io/mathlib4_docs/Mathlib/MeasureTheory/Decomposition/UnsignedHahn.html#MeasureTheory.hahn_decomposition), the only declaration that it contains) stumps a linter that looks for "collapsible" pairs `forall`/`intro`.

`have d_split` used to have a `MeasurableSet s` assumption, but it was unnecessary and I removed it.  Strictly speaking, this was not needed for my purpose, but I noticed it thanks to the linter.